# Conflicts:
#	jayud-common/src/main/java/com/jayud/common/enums/OrderTypeEnum.java
#	jayud-finance/jayud-finance-service/src/main/resources/com/jayud/finance/mapper/OrderPaymentBillMapper.xml
#	jayud-finance/jayud-finance-service/src/main/resources/com/jayud/finance/mapper/OrderReceivableBillMapper.xml

+++ jayud-oms/jayud-oms-model/src/main/java/com/jayud/oms/model/po/OrderReceivableCost.java
@@ -116,7 +116,7 @@ public class OrderReceivableCost extends Model<OrderReceivableCost> {
     private String legalName;
 
     @ApiModelProperty(value = "子订单法人主体id")
-    private Integer legalId;
+    private Long legalId;
 
     @Override
     protected Serializable pkVal() {
